requirements.txt
Django==5.0.4
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.3
Pillow==10.3.0